+++ src/components/Create/Parameters.tsx
@@ -57,7 +57,7 @@ const Parameters = ({
       categoricalVariables &&
       categoricalVariables
         .map(v => lookupCallback(v.code))
-        .filter(v => v.type === 'polynominal' || v.type === 'binominal');
+        .filter(v => v.type === 'multinominal' || v.type === 'binominal');
 
     const first = (vars && vars.length && vars[0]) || undefined;
     if (first && first.enumerations) {